+++ ARCPORTAL-SDK-APP/src/main/java/org/akraino/portal_user_interface/arcportalsdkapp/controller/BlueprintController.java
@@ -16,9 +16,9 @@
 
 package org.akraino.portal_user_interface.arcportalsdkapp.controller;
 
-import org.akraino.portal_user_interface.arcportalsdkapp.client.arc.resources.region.Region;
-import org.akraino.portal_user_interface.arcportalsdkapp.client.arc.resources.region.Regions;
-import org.akraino.portal_user_interface.arcportalsdkapp.service.RegionsService;
+import org.akraino.portal_user_interface.arcportalsdkapp.client.arc.resources.blueprint.Blueprint;
+import org.akraino.portal_user_interface.arcportalsdkapp.client.arc.resources.blueprint.Blueprints;
+import org.akraino.portal_user_interface.arcportalsdkapp.service.BlueprintService;
 import org.onap.portalsdk.core.controller.RestrictedBaseController;
 import org.onap.portalsdk.core.logging.logic.EELFLoggerDelegate;
 import org.onap.portalsdk.core.web.support.UserUtils;
@@ -27,41 +27,52 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping("/api/v1/region")
-public class RegionsController extends RestrictedBaseController {
+@RequestMapping("/api/v1/blueprint")
+public class BlueprintController extends RestrictedBaseController {
 
-    private static final EELFLoggerDelegate LOGGER = EELFLoggerDelegate.getLogger(RegionsController.class);
+    private static final EELFLoggerDelegate LOGGER = EELFLoggerDelegate.getLogger(BlueprintController.class);
 
     @Autowired
-    RegionsService service;
+    BlueprintService service;
 
-    public RegionsController() {
+    public BlueprintController() {
         super();
     }
 
     @RequestMapping(value = { "/" }, method = RequestMethod.GET)
-    public ResponseEntity<Regions> getRegions() {
+    public ResponseEntity<Blueprints> getBlueprints() {
         try {
-            return new ResponseEntity<>(service.getRegions(), HttpStatus.OK);
+            return new ResponseEntity<>(service.getBlueprints(), HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.error(EELFLoggerDelegate.errorLogger,
-                    "Error occured when trying to retrieve Edge Sites. " + UserUtils.getStackTrace(e));
+                    "Error occured when trying to retrieve Blueprints. " + UserUtils.getStackTrace(e));
         }
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
     }
 
     @RequestMapping(value = { "/{id}" }, method = RequestMethod.GET)
-    public ResponseEntity<Region> getRegion(@PathVariable("id") String uuid) {
+    public ResponseEntity<Blueprint> getBlueprint(@PathVariable("id") String uuid) {
         try {
-            return new ResponseEntity<>(service.getRegion(uuid), HttpStatus.OK);
+            return new ResponseEntity<>(service.getBlueprint(uuid), HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.error(EELFLoggerDelegate.errorLogger,
-                    "Error when retrieving Edge Site. " + UserUtils.getStackTrace(e));
+                    "Error when retrieving Blueprint. " + UserUtils.getStackTrace(e));
         }
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+
+    @RequestMapping(value = { "/" }, method = RequestMethod.POST)
+    public ResponseEntity<Blueprint> createBlueprint(@RequestBody String blueprint) {
+        try {
+            return new ResponseEntity<>(service.saveBlueprint(blueprint), HttpStatus.CREATED);
+        } catch (Exception e) {
+            LOGGER.error(EELFLoggerDelegate.errorLogger, "Creation of blueprint failed. " + UserUtils.getStackTrace(e));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+        }
     }
 }